Add exit and export buttons
