+++ pkg/model/service/servactive/service_constructor.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"strings"
 
+	"os"
+
 	"github.com/containerum/chkit/pkg/model/service"
 	"github.com/containerum/chkit/pkg/porta"
 	"github.com/containerum/chkit/pkg/util/activekit"
@@ -92,6 +94,13 @@ func Wizard(config ConstructorConfig) (service.Service, error) {
 							return nil
 						}
 						exit = true
+						return nil
+					},
+				},
+				{
+					Label: "Exit",
+					Action: func() error {
+						os.Exit(0)
 						return nil
 					},
 				},
